+++ RELEASES.md
@@ -25,6 +25,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: making getTypedSuccessor the defaul, used in service
 * service-templates-core: update to recursive traversal to visualise template instantiation, to support derivation, responsibility, specialization, membership 
 * service-template-core/archetype: form.html, visualisation of icons in navigation and documentation
+* service-template-core: backward_traversal in TemplateQuery and Template2Dot updated to support "overlay templates" without outputs.
 
 `2.2.3`
 * prov-template: update to instantiation